Add parent relative path to POM
Building the project with a clean repository fails
because the parent artifact cannot be found.

By adding the relative path to the parent
Maven will first search the path given for the project's parent,
before searching the local and then remote repositories.

As a result the parent artifact can be found and the build will succeed.

Signed-off-by: Davy Meers <davymeers@hotmail.com>